backend/requirements.txt
# FastAPI & Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Database
sqlalchemy==2.0.36
alembic==1.14.1
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Redis & Celery
redis==5.2.1
celery==5.4.0

# AI & RAG
langchain==0.3.14
langchain-google-genai==2.0.8
langchain-community==0.3.14
chromadb==0.5.23
google-generativeai==0.8.4

# HTTP & Webhooks
httpx==0.28.1
aiohttp==3.11.11

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.3.0

# Utilities
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
pillow==11.1.0
python-dateutil==2.9.0

# Auth
PyJWT==2.10.1

# CORS
starlette==0.41.3

# Logging
loguru==0.7.3
